#116 fix: 각종 오류들 수정
