+++ src/page/archiving/type/documentType.ts
@@ -1,7 +1,7 @@
 export interface DocumentType {
   documentId?: number;
   fileName: string;
-  fileUrl?: string;
+  fileUrl: string;
   blockName?: string;
   color?: string;
 }
