+++ src/shared/api/documents/team/document/index.ts
@@ -1,0 +1,7 @@
+import { axiosInstance } from '@/shared/api/instance';
+
+export const deleteDocument = async (teamId: number, documentId: number) => {
+  const response = await axiosInstance.delete(`/documents/team/${teamId}/document/${documentId}`);
+
+  return response;
+};
