+++ src/shared/api/time-blocks/team/time-block/index.ts
@@ -6,3 +6,9 @@ export const getDocuments = async (teamId: number, blockId: number) => {
 
   return response.data;
 };
+
+export const deleteBlock = async (teamId: number, timeBlockId: number) => {
+  const response = await axiosInstance.delete(`/time-blocks/team/${teamId}/time-block/${timeBlockId}`);
+
+  return response;
+};
